backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
qrcode==8.0
sqlparse==0.5.3
typing_extensions==4.12.2
validators==0.34.0
whitenoise==6.9.0
